src/requirements.txt
cloudshell-automation-api>=7.1.0,<7.2.0
cloudshell-shell-core>=3.1.0,<3.2.0
cloudshell-cli>=3.0.0,<3.1.0